+++ src/ui/components/Footer.js
@@ -1,4 +1,4 @@
-import Component from "./Component.js";
+import Component from "./core/Component.js";
 
 const Copyright = {
   YEAR: new Date().getFullYear(),
