+++ src/ui/components/core/Component.js
