+++ src/ui/components/game/LottoList.js
@@ -1,4 +1,4 @@
-import Component from "./Component.js";
+import Component from "../core/Component.js";
 
 class LottoList extends Component {
   template() {
