+++ src/ui/components/game/game.css
@@ -1,32 +1,3 @@
-#app {
-  display: flex;
-  flex-direction: column;
-  height: 100dvh;
-}
-
-#header {
-  background-color: var(--color-primary);
-  color: var(--color-greyscale-1);
-  padding: 14px 20px;
-}
-
-#main {
-  flex: 1;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-}
-
-#footer {
-  border-top: 1px solid var(--color-primary-opacity-20);
-  color: var(--color-primary);
-  padding-block: 28px;
-  display: flex;
-  p {
-    margin: auto;
-  }
-}
-
 /* LottoGame */
 #lotto-game {
   width: min(100%, 400px);
@@ -60,13 +31,6 @@
     padding: 8px;
     border-radius: 2px;
     flex: 1;
-  }
-
-  button {
-    background-color: var(--color-primary);
-    border-radius: 2px;
-    padding: 8px 10px;
-    color: var(--color-greyscale-1);
   }
 }
 
@@ -113,17 +77,6 @@
 }
 
 .open-result-button {
-  background-color: var(--color-primary);
-  color: var(--color-greyscale-1);
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  padding-block: 10px;
   width: 100%;
   margin-top: 20px;
-  border-radius: 5px;
-}
-.open-result-button:disabled {
-  opacity: 0.5;
-  cursor: not-allowed;
 }
