+++ src/ui/styles/layout.css
@@ -1,0 +1,28 @@
+#app {
+  display: flex;
+  flex-direction: column;
+  height: 100dvh;
+}
+
+#header {
+  background-color: var(--color-primary);
+  color: var(--color-greyscale-1);
+  padding: 14px 20px;
+}
+
+#main {
+  flex: 1;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+}
+
+#footer {
+  border-top: 1px solid var(--color-primary-opacity-20);
+  color: var(--color-primary);
+  padding-block: 28px;
+  display: flex;
+  p {
+    margin: auto;
+  }
+}
